+++ Sources/Swift/Simulation Extensions/UICollectionViewSimulation.swift
