+++ Sources/Swift/Simulation Extensions/UISegmentedControlSimulation.swift
