+++ Sources/Swift/Simulation Extensions/UISliderSimulation.swift
