+++ Sources/Swift/Simulation Extensions/UIStepperSimulation.swift
