+++ Sources/Swift/Simulation Extensions/UISwitchSimulation.swift
