+++ Sources/Swift/Simulation Extensions/UIWindowSimulation.swift
