+++ Sources/Swift/Utility Extensions/GeometryExtensions.swift
